refactor: split large service modules
