+++ band-platform/backend/app/api/role_management.py
@@ -9,7 +9,6 @@ import logging
 from typing import List, Optional, Dict
 
 from fastapi import APIRouter, Depends, HTTPException, BackgroundTasks
-from pydantic import BaseModel, Field, validator
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
@@ -20,119 +19,22 @@ from ..models.folder_structure import SyncStatus
 from ..services.content_parser import INSTRUMENT_KEY_MAPPING, get_keys_for_instruments
 from ..services.file_synchronizer import FileSynchronizer, schedule_sync_for_users
 from ..services.google_drive import GoogleDriveService
+from .role_models import (
+    InstrumentUpdate,
+    RoleUpdate,
+    AccessibleFilesResponse,
+    InstrumentReorganizeResponse,
+)
+from .role_helpers import (
+    get_current_user,
+    get_drive_credentials,
+    list_available_instruments,
+    list_available_roles,
+)
 
 logger = logging.getLogger(__name__)
 
 router = APIRouter()
-
-
-# TODO: Replace with actual authentication dependency
-async def get_current_user(
-    session: AsyncSession = Depends(get_db_session_dependency)
-) -> User:
-    """
-    Placeholder for getting the current authenticated user.
-    
-    This should be replaced with actual JWT/OAuth authentication.
-    """
-    # For now, return a mock user for testing
-    # In production, this would validate JWT token and return the actual user
-    result = await session.execute(
-        select(User)
-        .options(
-            selectinload(User.user_folder),
-            selectinload(User.band)
-        )
-        .where(User.id == 1)  # Mock user ID
-    )
-    user = result.scalar_one_or_none()
-    
-    if not user:
-        raise HTTPException(status_code=401, detail="User not authenticated")
-    
-    return user
-
-
-async def get_drive_credentials():
-    """
-    Get Google Drive credentials for the current user/band.
-    
-    This is a placeholder that should be replaced with actual credential management.
-    """
-    # TODO: Implement actual credential retrieval from database or OAuth flow
-    return None
-
-
-# Pydantic schemas for role management
-
-class InstrumentUpdate(BaseModel):
-    """Schema for updating user instruments."""
-    instruments: List[str] = Field(..., description="List of instrument names")
-    primary_instrument: Optional[str] = Field(None, description="Primary instrument")
-    reorganize_folders: bool = Field(
-        default=True, 
-        description="Whether to reorganize folders after instrument change"
-    )
-    
-    @validator('instruments')
-    def validate_instruments(cls, v):
-        """Validate that all instruments are recognized."""
-        if not v:
-            raise ValueError("At least one instrument must be specified")
-        
-        for instrument in v:
-            normalized = instrument.lower().replace(" ", "_").replace("-", "_")
-            if normalized not in INSTRUMENT_KEY_MAPPING:
-                raise ValueError(f"Unknown instrument: {instrument}")
-        
-        return v
-    
-    @validator('primary_instrument')
-    def validate_primary_instrument(cls, v, values):
-        """Validate that primary instrument is in the instruments list."""
-        if v and 'instruments' in values:
-            normalized_primary = v.lower().replace(" ", "_").replace("-", "_")
-            normalized_instruments = [
-                inst.lower().replace(" ", "_").replace("-", "_") 
-                for inst in values['instruments']
-            ]
-            
-            if normalized_primary not in normalized_instruments:
-                raise ValueError("Primary instrument must be in the instruments list")
-        
-        return v
-
-
-class RoleUpdate(BaseModel):
-    """Schema for updating user role."""
-    role: UserRole = Field(..., description="New user role")
-    reorganize_folders: bool = Field(
-        default=False,
-        description="Whether to reorganize folders after role change"
-    )
-
-
-class AccessibleFilesResponse(BaseModel):
-    """Response schema for accessible files query."""
-    user_id: int
-    instruments: List[str]
-    accessible_keys: List[str]
-    total_files: int
-    accessible_files: int
-    files_by_key: Dict[str, int]
-    files_by_type: Dict[str, int]
-
-
-class InstrumentReorganizeResponse(BaseModel):
-    """Response schema for instrument update with reorganization."""
-    user_id: int
-    old_instruments: List[str]
-    new_instruments: List[str]
-    old_keys: List[str]
-    new_keys: List[str]
-    reorganization_status: str
-    estimated_completion: Optional[str] = None
-    job_id: Optional[str] = None
 
 
 @router.put("/users/{user_id}/instruments", response_model=InstrumentReorganizeResponse, tags=["Role Management"])
@@ -514,102 +416,12 @@ async def get_user_accessible_files(
         )
 
 
+
 @router.get("/instruments", tags=["Role Management"])
-async def list_available_instruments():
-    """
-    List all available instruments with their transposition keys.
-    
-    Returns:
-        Dictionary of instruments organized by transposition key.
-    """
-    try:
-        # Group instruments by key
-        instruments_by_key = {}
-        for instrument, key in INSTRUMENT_KEY_MAPPING.items():
-            if key not in instruments_by_key:
-                instruments_by_key[key] = []
-            
-            # Convert to display format
-            display_name = instrument.replace("_", " ").title()
-            instruments_by_key[key].append({
-                "name": instrument,
-                "display_name": display_name,
-                "key": key
-            })
-        
-        # Sort instruments within each key group
-        for key in instruments_by_key:
-            instruments_by_key[key].sort(key=lambda x: x["display_name"])
-        
-        return {
-            "instruments_by_key": instruments_by_key,
-            "all_keys": sorted(instruments_by_key.keys()),
-            "total_instruments": len(INSTRUMENT_KEY_MAPPING),
-            "key_descriptions": {
-                "C": "Concert pitch instruments",
-                "Bb": "B-flat transposing instruments", 
-                "Eb": "E-flat transposing instruments",
-                "F": "F transposing instruments"
-            }
-        }
-        
-    except Exception as e:
-        logger.error(f"Error listing instruments: {e}")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to list instruments: {str(e)}"
-        )
+async def list_available_instruments_endpoint():
+    return await list_available_instruments()
 
 
 @router.get("/roles", tags=["Role Management"])
-async def list_available_roles():
-    """
-    List all available user roles and their descriptions.
-    
-    Returns:
-        Dictionary of roles with descriptions and permissions.
-    """
-    try:
-        roles_info = {
-            UserRole.MUSICIAN: {
-                "name": "Musician",
-                "description": "Regular band member with access to their instrument-specific content",
-                "permissions": [
-                    "View own folders",
-                    "Sync own folders",
-                    "Update own instruments"
-                ]
-            },
-            UserRole.BAND_LEADER: {
-                "name": "Band Leader",
-                "description": "Band leader with administrative privileges for the band",
-                "permissions": [
-                    "All musician permissions",
-                    "View all band members' folders",
-                    "Manage band members' roles and instruments",
-                    "Manage band Google Drive integration"
-                ]
-            },
-            UserRole.ADMIN: {
-                "name": "Administrator",
-                "description": "Platform administrator with full access",
-                "permissions": [
-                    "All band leader permissions",
-                    "Manage multiple bands",
-                    "View platform statistics",
-                    "Manage system settings"
-                ]
-            }
-        }
-        
-        return {
-            "roles": roles_info,
-            "role_hierarchy": [UserRole.MUSICIAN, UserRole.BAND_LEADER, UserRole.ADMIN]
-        }
-        
-    except Exception as e:
-        logger.error(f"Error listing roles: {e}")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to list roles: {str(e)}"
-        )
+async def list_available_roles_endpoint():
+    return await list_available_roles()
