+++ band-platform/backend/app/services/google_drive.py
@@ -15,6 +15,14 @@ from contextlib import asynccontextmanager
 from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import Flow
+
+from .drive_helpers import (
+    RateLimiter,
+    GoogleDriveAuth,
+    StatsMixin,
+    create_drive_service,
+    test_drive_connection,
+)
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 import backoff
@@ -40,7 +48,7 @@ class AuthenticationError(DriveAPIError):
     pass
 
 
-class GoogleDriveService:
+class GoogleDriveService(StatsMixin):
     """
     Google Drive API service with comprehensive error handling and rate limiting.
     
@@ -60,13 +68,7 @@ class GoogleDriveService:
         self.rate_limiter = RateLimiter(requests_per_second=10)  # Google Drive limit: 1000/100s
         
         # Statistics
-        self.stats = {
-            "requests_made": 0,
-            "rate_limit_hits": 0,
-            "errors": 0,
-            "files_processed": 0,
-            "last_sync": None,
-        }
+        self.init_stats()
     
     @asynccontextmanager
     async def _get_service(self):
@@ -485,153 +487,9 @@ class GoogleDriveService:
     
     def get_stats(self) -> Dict[str, Any]:
         """Get service statistics."""
-        return self.stats.copy()
-    
+        return super().get_stats()
+
     def reset_stats(self) -> None:
         """Reset service statistics."""
-        self.stats = {
-            "requests_made": 0,
-            "rate_limit_hits": 0,
-            "errors": 0,
-            "files_processed": 0,
-            "last_sync": None,
-        }
+        super().reset_stats()
 
-
-class RateLimiter:
-    """
-    Rate limiter for Google API requests.
-    
-    Implements token bucket algorithm for smooth rate limiting.
-    """
-    
-    def __init__(self, requests_per_second: float):
-        self.requests_per_second = requests_per_second
-        self.tokens = requests_per_second
-        self.last_update = asyncio.get_event_loop().time()
-        self.lock = asyncio.Lock()
-    
-    async def acquire(self):
-        """Acquire a rate limit token."""
-        async with self.lock:
-            now = asyncio.get_event_loop().time()
-            time_passed = now - self.last_update
-            
-            # Add tokens based on time passed
-            self.tokens = min(
-                self.requests_per_second,
-                self.tokens + time_passed * self.requests_per_second
-            )
-            self.last_update = now
-            
-            if self.tokens >= 1:
-                self.tokens -= 1
-                return
-            
-            # Need to wait
-            wait_time = (1 - self.tokens) / self.requests_per_second
-            await asyncio.sleep(wait_time)
-            self.tokens = 0
-
-
-class GoogleDriveAuth:
-    """
-    Google Drive authentication manager.
-    
-    Handles OAuth 2.0 flow and credential management.
-    """
-    
-    @staticmethod
-    def create_oauth_flow(redirect_uri: str) -> Flow:
-        """
-        Create OAuth 2.0 flow for Google Drive.
-        
-        Args:
-            redirect_uri: OAuth redirect URI.
-            
-        Returns:
-            Configured OAuth flow.
-        """
-        return Flow.from_client_config(
-            {
-                "web": {
-                    "client_id": settings.google_client_id,
-                    "client_secret": settings.google_client_secret,
-                    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-                    "token_uri": "https://oauth2.googleapis.com/token",
-                }
-            },
-            scopes=[settings.google_drive_scope],
-            redirect_uri=redirect_uri
-        )
-    
-    @staticmethod
-    def get_authorization_url(flow: Flow) -> Tuple[str, str]:
-        """
-        Get authorization URL and state for OAuth flow.
-        
-        Args:
-            flow: OAuth flow instance.
-            
-        Returns:
-            Tuple of (authorization_url, state).
-        """
-        authorization_url, state = flow.authorization_url(
-            access_type='offline',
-            include_granted_scopes='true',
-            prompt='consent'  # Force consent to get refresh token
-        )
-        return authorization_url, state
-    
-    @staticmethod
-    def exchange_code_for_credentials(flow: Flow, authorization_response: str) -> Credentials:
-        """
-        Exchange authorization code for credentials.
-        
-        Args:
-            flow: OAuth flow instance.
-            authorization_response: Full authorization response URL.
-            
-        Returns:
-            Google OAuth credentials.
-        """
-        flow.fetch_token(authorization_response=authorization_response)
-        return flow.credentials
-
-
-# Convenience functions for service creation
-
-def create_drive_service(credentials: Credentials) -> GoogleDriveService:
-    """
-    Create a Google Drive service instance.
-    
-    Args:
-        credentials: Google OAuth credentials.
-        
-    Returns:
-        Configured GoogleDriveService instance.
-    """
-    return GoogleDriveService(credentials)
-
-
-async def test_drive_connection(service: GoogleDriveService) -> bool:
-    """
-    Test Google Drive API connection.
-    
-    Args:
-        service: GoogleDriveService instance.
-        
-    Returns:
-        True if connection is successful.
-    """
-    try:
-        # Try to get user info
-        async with service._get_service() as drive:
-            def make_request():
-                return drive.about().get(fields="user").execute()
-            
-            await service._make_request(make_request)
-            return True
-    except Exception as e:
-        logger.error(f"Drive connection test failed: {e}")
-        return False
